+++ src/options/pages/blueprints/modals/ExtensionLogsModal.module.scss
