+++ src/options/pages/blueprints/modals/ShareExtensionModal.module.scss
